editRatio {editRatio}	R Documentation

Calculate the edit distance ratio for phonetic transcriptions

Description

editRatio returns the edit distance ratio comparing the target transcription(s) to the produced transcription(s).

Usage


editRatio(x)

Arguments

x
a data frame that contains target transcription and produced transcription on alternating rows

Value

The edit distance ratio, the number of edits (Levenshtein, 1966) divided by the number of target or produced segments, whichever is greater.

References

Levenshtein, V. I. (1966). Binary codes capable of correcting deletions, insertions and reversals. Soviet Physics Doklady, 10 (8).

Examples


x<- data.frame("Transcription"=c("Saul", "Paul"))
editRatio(x)


[Package editRatio version 0.1.0 Index]